+++ demisto_sdk/commands/init/templates/HelloWorldScript/HelloWorldScript.py
@@ -1,6 +1,6 @@
 import demistomock as demisto
-from CommonServerPython import *
-from CommonServerUserPython import *
+from CommonServerPython import *  # noqa: E402 lgtm [py/polluting-import]
+from CommonServerUserPython import *  # noqa: E402 lgtm [py/polluting-import]
 
 
 def say_hello(name):